feat(pkg/*): change IsPermissiveTrafficPolicyMode to check if the provided identity is for the gateway
Add the serviceIdentity to every call of IsPermissiveTrafficPolicyMode.
Move some of the utils Serving functionality to pkg/envoy/ads to avoid import cycles.
