+++ pkg/envoy/ads/stream.go
@@ -8,6 +8,10 @@ import (
 	mapset "github.com/deckarep/golang-set"
 	xds_discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
 	"github.com/pkg/errors"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/credentials"
+	"google.golang.org/grpc/peer"
+	"google.golang.org/grpc/status"
 
 	"github.com/openservicemesh/osm/pkg/announcements"
 	"github.com/openservicemesh/osm/pkg/certificate"
@@ -24,7 +28,7 @@ import (
 func (s *Server) StreamAggregatedResources(server xds_discovery.AggregatedDiscoveryService_StreamAggregatedResourcesServer) error {
 	// When a new Envoy proxy connects, ValidateClient would ensure that it has a valid certificate,
 	// and the Subject CN is in the allowedCommonNames set.
-	certCommonName, certSerialNumber, err := utils.ValidateClient(server.Context(), nil)
+	certCommonName, certSerialNumber, err := ValidateClient(server.Context(), nil)
 	if err != nil {
 		return errors.Wrap(err, "Could not start Aggregated Discovery Service gRPC stream for newly connected Envoy proxy")
 	}
@@ -351,4 +355,34 @@ func (s *Server) recordPodMetadata(p *envoy.Proxy) error {
 	}
 
 	return nil
+}
+
+// ValidateClient ensures that the connected client is authorized to connect to the gRPC server.
+func ValidateClient(ctx context.Context, allowedCommonNames map[string]interface{}) (certificate.CommonName, certificate.SerialNumber, error) {
+	mtlsPeer, ok := peer.FromContext(ctx)
+	if !ok {
+		log.Error().Msg("[grpc][mTLS] No peer found")
+		return "", "", status.Error(codes.Unauthenticated, "no peer found")
+	}
+
+	tlsAuth, ok := mtlsPeer.AuthInfo.(credentials.TLSInfo)
+	if !ok {
+		log.Error().Msg("[grpc][mTLS] Unexpected peer transport credentials")
+		return "", "", status.Error(codes.Unauthenticated, "unexpected peer transport credentials")
+	}
+
+	if len(tlsAuth.State.VerifiedChains) == 0 || len(tlsAuth.State.VerifiedChains[0]) == 0 {
+		log.Error().Msgf("[grpc][mTLS] Could not verify peer certificate")
+		return "", "", status.Error(codes.Unauthenticated, "could not verify peer certificate")
+	}
+
+	// Check whether the subject common name is one that is allowed to connect.
+	cn := tlsAuth.State.VerifiedChains[0][0].Subject.CommonName
+	if _, ok := allowedCommonNames[cn]; len(allowedCommonNames) > 0 && !ok {
+		log.Error().Msgf("[grpc][mTLS] Subject common name %+v not allowed", cn)
+		return "", "", status.Error(codes.Unauthenticated, "disallowed subject common name")
+	}
+
+	certificateSerialNumber := tlsAuth.State.VerifiedChains[0][0].SerialNumber.String()
+	return certificate.CommonName(cn), certificate.SerialNumber(certificateSerialNumber), nil
 }
